// EigenDA/EigenDA-1/src/types/schema.ts
// THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.

import { Document } from "@blockflow-labs/utils";

export class Batch {
  static entity = "Batch";
  static schema = {
    id: { type: "String", index: true },
    batch_id: "string",
    batch_header_hash: "string",
    batch_header: "string",
    non_signing: "string",
    gas_fees: "string",
    block_timestamp: "string",
    transaction_hash: "string",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}

export class GasFees {
  static entity = "GasFees";
  static schema = {
    id: { type: "String", index: true },
    gas_used: "string",
    gas_price: "string",
    transaction_fee: "string",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}

export class BatchHeader {
  static entity = "BatchHeader";
  static schema = {
    id: { type: "String", index: true },
    blob_headers_root: "string",
    quorum_numbers: ["string"],
    signed_stake_for_quorums: ["string"],
    reference_blocknumber: "string",
    batch: "string",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}

export class NonSigning {
  static entity = "NonSigning";
  static schema = {
    id: { type: "String", index: true },
    non_signers: ["string"],
    batch: "string",
    signatory_record_hash: "string",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}

export class Operator {
  static entity = "Operator";
  static schema = {
    id: { type: "String", index: true },
    operator_id: "string",
    non_signings: ["string"],
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}

import { String } from "@blockflow-labs/utils";

export interface IBatch extends Document {
  id: String;
  batch_id: string;
  batch_header_hash: string;
  batch_header: string;
  non_signing: string;
  gas_fees: string;
  block_timestamp: string;
  transaction_hash: string;
  blocknumber: String;
  entityId: String;
  instanceId: String;
  chainId: String;
}

export interface IGasFees extends Document {
  id: String;
  gas_used: string;
  gas_price: string;
  transaction_fee: string;
  blocknumber: String;
  entityId: String;
  instanceId: String;
  chainId: String;
}

export interface IBatchHeader extends Document {
  id: String;
  blob_headers_root: string;
  quorum_numbers: [string];
  signed_stake_for_quorums: [string];
  reference_blocknumber: string;
  batch: string;
  blocknumber: String;
  entityId: String;
  instanceId: String;
  chainId: String;
}

export interface INonSigning extends Document {
  id: String;
  non_signers: [string];
  batch: string;
  signatory_record_hash: string;
  blocknumber: String;
  entityId: String;
  instanceId: String;
  chainId: String;
}

export interface IOperator extends Document {
  id: string;
  operator_id: string;
  non_signings: [string];
  blocknumber: String;
  entityId: String;
  instanceId: String;
  chainId: String;
}